pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentcore"
version = "0.1.0"
description = "A hot-loadable, markdown-first, codebase-aware role mesh for SDLC agent orchestration."
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "Apache-2.0" }
authors = [{ name = "agent-core-orchestrator contributors" }]
keywords = ["agents", "orchestration", "rag", "sdlc", "fastapi", "pgvector"]

dependencies = [
  # Core
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-frontmatter>=1.1",
  "pyyaml>=6.0",
  "watchfiles>=0.21",
  "structlog>=24.1",
  "typer>=0.12",
  "rich>=13.7",
  "httpx>=0.27",

  # Orchestrator
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",

  # LLM providers (kept thin; users opt in)
  "anthropic>=0.34",
  "openai>=1.30",            # Azure OpenAI + z.ai (OpenAI-compatible)
  "boto3>=1.34",             # Bedrock

  # Memory / RAG
  "psycopg[binary]>=3.1",
  "pgvector>=0.2.5",
  "numpy>=1.26",
  "networkx>=3.2",
  "python-louvain>=0.16",

  # Code indexing
  "tree-sitter>=0.21",
  "tree-sitter-languages>=1.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "mypy>=1.10",
  "respx>=0.21",
]
ollama = ["ollama>=0.2"]

[project.scripts]
agentcore = "agentcore.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/agentcore"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# uv-friendly: lockfile lives at uv.lock once you run `uv sync`.
# `uv run agentcore <cmd>` works out of the box with this pyproject.
[tool.uv]
package = true
dev-dependencies = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "mypy>=1.10",
  "respx>=0.21",
]
